+++ pkg/client/config_test.go
@@ -43,6 +43,8 @@ images:
   webhookAgentImage: ambassador-telepresence-webhook-image:0.0.2
 telepresenceAPI:
   port: 1234
+intercept:
+  appProtocolStrategy: portName
 `,
 	}
 
@@ -80,6 +82,7 @@ telepresenceAPI:
 	assert.Equal(t, "ambassador-telepresence-client-image:0.0.1", cfg.Images.AgentImage)         // from user
 	assert.Equal(t, "ambassador-telepresence-webhook-image:0.0.2", cfg.Images.WebhookAgentImage) // from user
 	assert.Equal(t, 1234, cfg.TelepresenceAPI.Port)                                              // from user
+	assert.Equal(t, PortName, cfg.Intercept.AppProtocolStrategy)                                 // from user
 }
 
 func Test_ConfigMarshalYAML(t *testing.T) {
@@ -94,6 +97,7 @@ func Test_ConfigMarshalYAML(t *testing.T) {
 	cfg.LogLevels.UserDaemon = logrus.TraceLevel
 	cfg.Grpc.MaxReceiveSize, _ = resource.ParseQuantity("20Mi")
 	cfg.TelepresenceAPI.Port = 4567
+	cfg.Intercept.AppProtocolStrategy = PortName
 	cfgBytes, err := yaml.Marshal(cfg)
 	require.NoError(t, err)