+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### 2.4.10 (TBD)
 
+- Feature: The strategy when selecting the application protocol for personal intercepts can now be configured using
+  the `intercept.appProtocolStrategy` in the `config.yml` file.
+
 - Bugfix: Telepresence will no longer log invalid: "unhandled connection control message: code DIAL_OK" errors.
 
 - Bugfix: User will not be asked to log in or add ingress information when creating an intercept until a check has been